convert set-ad-test-cookie to TS
